requirements.txt
Flask

gunicorn

requests

beautifulsoup4

newspaper3k